i think it's finish sidebar
