+++ src/app/blog/page.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const BlogPage = () => {
+  return (
+    <div>BlogPage</div>
+  )
+}
+
+export default BlogPage
