+++ src/app/projects/page.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ProjectPage = () => {
+  return (
+    <div>ProjectPage</div>
+  )
+}
+
+export default ProjectPage
