+++ src/components/test.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Test = () => {
+  return (
+    <div className='flex justify-center items-center text-5xl'>
+        This is test
+    </div>
+  )
+}
+
+export default Test
